Fixed image positions for different screens
Before, the clouds inside the top container of version2.0 was positioned using absolute attribute. 

Due to this, if a user's browser was zoomed it or out, the clouds would move accordingly. This lead to cases where the clouds would block the header and intro, or cases where the clouds were just too far spread out. 

In version2.0.1, no content has changed. Only the way the elements inside the top container are laid out has changed.